+++ src/main/java/module-info.java
@@ -1,4 +1,8 @@
 module org.openjfx {
     requires javafx.controls;
-    exports org.openjfx;
+    requires javafx.fxml;
+    requires javafx.base;
+    requires javafx.graphics;
+    exports net.xz3ra.www.karaoke;
+    //exports net.xz3ra.www.karaoke.components;
 }
